# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/camps/{campId}/groups": {
            "get": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "특정 캠프에 속한 모든 조의 목록과 상태를 조회한다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "B. Resource Management (Admin)"
                ],
                "summary": "전체 조 목록 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "캠프 ID",
                        "name": "campId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_infrastructure_web.Group"
                            }
                        }
                    }
                }
            }
        },
...
        "/groups/{id}": {
            "get": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "특정 조의 현재 위치 및 순회표(Itinerary) 진행 상태를 조회한다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "B. Resource Management (Admin)"
                ],
                "summary": "특정 조 상세 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "조 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web.Group"
                        }
                    }
                }
            }
        },
        "/groups/{id}/visits": {
            "get": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "특정 조의 전체 방문(Visit) 기록과 각 코너의 소요 시간 등을 조회한다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "B. Resource Management (Admin)"
                ],
                "summary": "조별 방문 기록 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "조 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_infrastructure_web.VisitSummary"
                            }
                        }
                    }
                }
            }
        },
        "/messages/broadcast": {
            "get": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "관리자가 보낸 BROADCAST 메시지들의 목록을 조회한다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "E. Message"
                ],
                "summary": "발송된 공지사항 목록",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_infrastructure_web.Message"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "모든 활성 트랙에 BROADCAST 메시지를 보낸다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "E. Message"
                ],
                "summary": "전체 공지 발송",
                "parameters": [
                    {
                        "description": "메시지 내용",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web.BroadcastMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web.Message"
                        }
                    }
                }
            }
        },
        "/messages/broadcast/{id}/read": {
            "post": {
                "security": [
                    {
                        "TrackAuth": []
                    }
                ],
                "description": "트랙 진행자가 공지사항을 확인(읽음) 처리한다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "E. Message"
                ],
                "summary": "공지사항 읽음 처리",
                "parameters": [
                    {
                        "type": "string",
                        "description": "메시지 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "성공적으로 읽음 처리됨"
                    }
                }
            }
        },
        "/messages/broadcast/{id}/receipts": {
            "get": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "특정 공지사항에 대해 트랙들의 수신/읽음 상태를 확인한다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "E. Message"
                ],
                "summary": "공지사항 수신 확인 현황",
                "parameters": [
                    {
                        "type": "string",
                        "description": "메시지 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_infrastructure_web.BroadcastReceipt"
                            }
                        }
                    }
                }
            }
        },
...
        "/tracks/{trackId}/messages": {
            "get": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "관리자 또는 트랙 진행자가 해당 트랙과 관련된 DIRECT 메시지 내역을 조회한다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "E. Message"
                ],
                "summary": "트랙별 메시지 내역 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "트랙 ID",
                        "name": "trackId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_infrastructure_web.Message"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "관리자가 특정 트랙에, 또는 특정 트랙이 관리자에게 DIRECT 메시지를 발송한다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "E. Message"
                ],
                "summary": "다이렉트 메시지 발송",
                "parameters": [
                    {
                        "type": "string",
                        "description": "트랙 ID",
                        "name": "trackId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "메시지 내용",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web.DirectMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web.Message"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_infrastructure_web.BroadcastMessageRequest": {
            "type": "object",
            "properties": {
                "campId": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                }
            }
        },
        "internal_infrastructure_web.BroadcastReceipt": {
            "type": "object",
            "properties": {
                "cornerName": {
                    "type": "string"
                },
                "isRead": {
                    "type": "boolean"
                },
                "readAt": {
                    "type": "string",
                    "format": "date-time"
                },
                "trackId": {
                    "type": "string",
                    "format": "uuid"
                },
                "trackNo": {
                    "type": "integer"
                }
            }
        },
...
        "internal_infrastructure_web.CornerProgress": {
            "type": "object",
            "properties": {
                "cornerId": {
                    "type": "string",
                    "format": "uuid"
                },
                "cornerName": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "NOT_VISITED",
                        "IN_PROGRESS",
                        "COMPLETED"
                    ]
                }
            }
        },
...
        "internal_infrastructure_web.DirectMessageRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                }
            }
        },
...
        "internal_infrastructure_web.Group": {
            "type": "object",
            "properties": {
                "badgeId": {
                    "type": "string",
                    "format": "uuid"
                },
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "isFinished": {
                    "type": "boolean"
                },
                "itinerary": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_infrastructure_web.CornerProgress"
                    }
                },
                "name": {
                    "type": "string",
                    "example": "1조"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "IDLE_MOVING",
                        "AT_CORNER",
                        "FINISHED"
                    ]
                }
            }
        },
...
        "internal_infrastructure_web.Message": {
            "type": "object",
            "properties": {
                "channelType": {
                    "type": "string",
                    "enum": [
                        "BROADCAST",
                        "DIRECT"
                    ]
                },
                "content": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "isRead": {
                    "type": "boolean"
                },
                "readAt": {
                    "type": "string",
                    "format": "date-time"
                },
                "senderRole": {
                    "type": "string",
                    "enum": [
                        "ADMIN",
                        "TRACK"
                    ]
                },
                "sentAt": {
                    "type": "string",
                    "format": "date-time"
                },
                "trackId": {
                    "type": "string",
                    "format": "uuid"
                }
            }
        },
...
        "internal_infrastructure_web.VisitSummary": {
            "type": "object",
            "properties": {
                "cornerId": {
                    "type": "string",
                    "format": "uuid"
                },
                "deviationSeconds": {
                    "type": "integer"
                },
                "durationSeconds": {
                    "type": "integer"
                },
                "endedAt": {
                    "type": "string",
                    "format": "date-time"
                },
                "groupId": {
                    "type": "string",
                    "format": "uuid"
                },
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "inputMethod": {
                    "type": "string",
                    "enum": [
                        "QR_SCAN",
                        "MANUAL"
                    ]
                },
                "startedAt": {
                    "type": "string",
                    "format": "date-time"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "IN_PROGRESS",
                        "COMPLETED"
                    ]
                },
                "trackId": {
                    "type": "string",
                    "format": "uuid"
                }
            }
        }
    }
}